+++ stretch-core/src/node.rs
@@ -14,17 +14,17 @@ pub enum NeuronType {
 pub struct Node {
     pub id: usize,
     /// Niveau d'activité courant [0, ∞) borné en pratique
-    pub activation: f64,
+    pub activation: f32,
     /// Seuil de déclenchement
-    pub threshold: f64,
+    pub threshold: f32,
     /// Fatigue / réfractarité courante
-    pub fatigue: f64,
+    pub fatigue: f32,
     /// Trace mémoire locale cumulée
-    pub memory_trace: f64,
+    pub memory_trace: f32,
     /// Facilité à répondre (multiplicateur)
-    pub excitability: f64,
+    pub excitability: f32,
     /// Résistance locale à l'activation
-    pub inhibition: f64,
+    pub inhibition: f32,
     /// Compteur d'activations (instrumentation)
     pub activation_count: u64,
     /// V3 : type de neurone (E/I)
@@ -32,19 +32,19 @@ pub struct Node {
     /// V3 : dernier tick où le nœud était actif (pour STDP)
     pub last_activation_tick: Option<usize>,
     /// V3 : modulation du seuil par le PID indirect de zone
-    pub threshold_mod: f64,
+    pub threshold_mod: f32,
 }
 
 impl Node {
     pub fn new(id: usize, defaults: &NodeDefaults) -> Self {
         Node {
             id,
-            activation: defaults.activation,
-            threshold: defaults.threshold,
-            fatigue: defaults.fatigue,
-            memory_trace: defaults.memory_trace,
-            excitability: defaults.excitability,
-            inhibition: defaults.inhibition,
+            activation: defaults.activation as f32,
+            threshold: defaults.threshold as f32,
+            fatigue: defaults.fatigue as f32,
+            memory_trace: defaults.memory_trace as f32,
+            excitability: defaults.excitability as f32,
+            inhibition: defaults.inhibition as f32,
             activation_count: 0,
             node_type: NeuronType::Excitatory,
             last_activation_tick: None,
@@ -54,7 +54,7 @@ impl Node {
 
     /// Seuil effectif corrigé par fatigue, excitabilité, inhibition et modulation de zone.
     /// Borné inférieurement à 0.01 pour éviter que le PID rende tous les nœuds actifs.
-    pub fn effective_threshold(&self) -> f64 {
+    pub fn effective_threshold(&self) -> f32 {
         let raw = (self.threshold + self.fatigue + self.inhibition + self.threshold_mod) / self.excitability.max(0.01);
         raw.max(0.05)
     }
@@ -67,13 +67,13 @@ impl Node {
     /// Appliquer la dissipation de l'activation avec potentiel de repos.
     /// `rate` : taux de decay (déjà jitté si applicable)
     /// `min`  : activation minimale (potentiel de repos)
-    pub fn decay_activation(&mut self, rate: f64, min: f64) {
+    pub fn decay_activation(&mut self, rate: f32, min: f32) {
         self.activation *= 1.0 - rate;
         self.activation = self.activation.max(min);
     }
 
     /// Mettre à jour la fatigue.
-    pub fn update_fatigue(&mut self, gain: f64, recovery: f64) {
+    pub fn update_fatigue(&mut self, gain: f32, recovery: f32) {
         if self.is_active() {
             self.fatigue += gain * self.activation;
         }
@@ -82,7 +82,7 @@ impl Node {
     }
 
     /// Mettre à jour l'inhibition.
-    pub fn update_inhibition(&mut self, gain: f64, decay: f64) {
+    pub fn update_inhibition(&mut self, gain: f32, decay: f32) {
         if self.is_active() {
             self.inhibition += gain;
         }
@@ -91,7 +91,7 @@ impl Node {
     }
 
     /// Mettre à jour la trace mémoire locale.
-    pub fn update_trace(&mut self, trace_gain: f64, trace_decay: f64) {
+    pub fn update_trace(&mut self, trace_gain: f32, trace_decay: f32) {
         if self.is_active() {
             self.memory_trace += trace_gain * self.activation;
             self.activation_count += 1;
@@ -107,7 +107,7 @@ impl Node {
     }
 
     /// Injecter un stimulus externe.
-    pub fn inject_stimulus(&mut self, intensity: f64) {
+    pub fn inject_stimulus(&mut self, intensity: f32) {
         self.activation += intensity;
     }
 }
